refactor: 변수, 컴포넌트명 수정 및 onClick 이벤트에 직접 함수 정의
- InputBox -> InputWithError 로 컴포넌트명 수정
- SelectBox -> Select 로 컴포넌트명 수정
- imageList -> filteredImgList 로 변수명 수정
- handleImageClick, handleProfileImgClick 함수를 onClick 이벤트에서 직접
  정의
