+++ src/pages/message/ProfileImg.jsx
@@ -1,22 +1,12 @@
 const ProfileImg = ({ images, profileImg, setProfileImg }) => {
-  // 선택된 예시 이미지를 프로필 이미지로 설정
-  const handleImageClick = (e) => {
-    setProfileImg(e.target.currentSrc);
-  };
-
-  // 선택된 프로필 이미지 제거
-  const handleProfileImgClick = (e) => {
-    setProfileImg(images[0]);
-  };
-
-  const imageList = images
+  const filteredImgList = images
     .filter((_, index) => index >= 1) // index 0 제외
     .map((src) => (
       <img
         key={src}
         src={src}
         className="w-14 h-14 rounded-full mr-1 cursor-pointer"
-        onClick={handleImageClick}
+        onClick={(e) => setProfileImg(e.target.currentSrc)} // 선택된 예시 이미지를 프로필로 지정
       />
     ));
   return (
@@ -24,11 +14,11 @@ const ProfileImg = ({ images, profileImg, setProfileImg }) => {
       <img
         src={profileImg}
         className="mr-8 w-20 h-20 rounded-full cursor-pointer"
-        onClick={handleProfileImgClick}
+        onClick={(e) => setProfileImg(images[0])} // 클릭 시 프로필 이미지 제거
       />
       <div>
         <p className="text-[#555555] mb-3">프로필 이미지를 선택해주세요!</p>
-        <div className="flex">{imageList}</div>
+        <div className="flex">{filteredImgList}</div>
       </div>
     </div>
   );
